move stage module
