+++ src/snowflake/cli/plugins/object/__init__.py
@@ -1,6 +1,6 @@
 from snowflake.cli.api.commands.snow_typer import SnowTyper
 from snowflake.cli.plugins.object.commands import app as show_app
-from snowflake.cli.plugins.object.stage.commands import app as stage_app
+from snowflake.cli.plugins.stage.commands import app as stage_app
 
 app = SnowTyper(
     name="object",
