+++ src/snowflake/cli/plugins/object/commands.py
@@ -10,7 +10,7 @@ from snowflake.cli.api.constants import SUPPORTED_OBJECTS, VALID_SCOPES
 from snowflake.cli.api.output.types import QueryResult
 from snowflake.cli.api.project.util import is_valid_identifier
 from snowflake.cli.plugins.object.manager import ObjectManager
-from snowflake.cli.plugins.object.stage.commands import app as stage_app
+from snowflake.cli.plugins.stage.commands import app as stage_app
 
 app = SnowTyper(
     name="object",
