+++ src/snowflake/cli/plugins/stage/diff.py
