+++ tests/stage/test_diff.py
@@ -5,7 +5,7 @@ from unittest import mock
 
 import pytest
 from snowflake.cli.api.exceptions import SnowflakeSQLExecutionError
-from snowflake.cli.plugins.object.stage.diff import (
+from snowflake.cli.plugins.stage.diff import (
     DiffResult,
     delete_only_on_stage_files,
     enumerate_files,
@@ -14,11 +14,11 @@ from snowflake.cli.plugins.object.stage.diff import (
     stage_diff,
     sync_local_diff_with_stage,
 )
-from snowflake.cli.plugins.object.stage.manager import StageManager
+from snowflake.cli.plugins.stage.manager import StageManager
 
 from tests.testing_utils.files_and_dirs import temp_local_dir
 
-STAGE_MANAGER = "snowflake.cli.plugins.object.stage.manager.StageManager"
+STAGE_MANAGER = "snowflake.cli.plugins.stage.manager.StageManager"
 
 FILE_CONTENTS = {
     "README.md": "This is a README\n",
